requirements.txt
asgiref==3.8.1
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.12.3
pytz==2024.2
setuptools==75.3.0
sqlparse==0.2.4
